+++ src/main/java/buildcraft/transport/stripes/StripesHandlerMinecartDestroy.java
@@ -9,7 +9,11 @@ import java.util.Collections;
 import java.util.List;
 
 import net.minecraft.world.entity.player.Player;
+//? if >=1.21.11 {
 import net.minecraft.world.entity.vehicle.minecart.AbstractMinecart;
+//?} else {
+/*import net.minecraft.world.entity.vehicle.AbstractMinecart;*/
+//?}
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.core.Direction;
 import net.minecraft.core.BlockPos;